Adjust file structure to match ansible-core
