+++ ansible.cfg
@@ -1,0 +1,5 @@
+[defaults]
+
+library = ./library
+module_utils   = ./module_utils
+
